Move LoggableGenerator to Generator subfolder
